Avoid losing ordering by using Lists instead of Sets
The JerseyWhiteboardDispatcher is responsible for disambiguating resources with clashing paths/names, however it has been using HashSet to collect things. This loses the sort ordering that is applied to the entries, and results in unstable behaviour. Using Lists will keep the sort ordering, at the expense of also keeping duplicates. This should not be a problem as none of the entries implement equals, and so are unlikely to be duplicates.

This fixes the TCK test testResourcesWithClashingNames

Signed-off-by: Tim Ward <timothyjward@apache.org>