Support proxying of extension types
Necessary to support registering more than one instance of the same type. ASM must be used to copy generics information and annotations accross to the proxy class. Proxy classes have names generated using the service id and ranking so that the type names sort according to service ranking order. A new instance must be therefore be registered if the extension changes its ranking.

Signed-off-by: Tim Ward <timothyjward@apache.org>